Convert setup and migrate to TS
